pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bases-athle-scrapper"
version = "0.1.0"
description = "Tools to scrape athletics clubs and athletes from bases.athle"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "Bases Athle Maintainers"}]
dependencies = [
    "beautifulsoup4",
    "lxml",
    "psycopg2-binary",
    "python-dotenv",
    "requests",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
